\name{ptf.macro.thetaSMi}
\alias{ptf.macro.thetaSMi}
\title{bla bla bla bla}
\description{bla bla bla (i}
\usage{
ptf.macro.thetaSMi(clay)
}
\arguments{
  \item{clay}{Vector of numericals. Clay content [\%] of each soil layer / 
horizon. }
}

\value{The function returns a vector of thetaSMi values of the same 
length as the vector of value provided to each parameter. 
Unit of thetaSMi is [-] or [m3 of water.m-3 of bulk soil].}
\references{Jarvis N., Lindahl, A., Messing, I., Stenemo, F., 
Hollis, J., Reichenberger, S. & Dubus, I.G., 2007. Algorithm 
to completely parameterise MACRO from basic soil property data. 
Report DL21 of the FP6 EU-funded FOOTPRINT project 
[www.eu-footprint.org], 18p.

Wosten J.H.M., Lilly A., Nemes A., 
Le Bas C., 1999. Development and use of a database of hydraulic 
properties of European soils. Geoderma 90:169-185.}
\author{tdh <toby.hocking@irnia.fr>}



\seealso{\code{\link{ptf.macro.prefFlow}}, 
\code{\link{ptf.macro}}.}

