\name{test-package}
\alias{test-package}
\docType{package}
\title{foooo}
\description{barrrrrrrrrrr}
\details{
\tabular{ll}{Package: \tab test\cr
Maintainer: \tab tdh <toby.hocking@irnia.fr>\cr
Author: \tab tdh\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab foooo\cr}
}
\author{tdh <toby.hocking@irnia.fr>}

\keyword{ package }


